requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
discord.py==2.5.2
frozenlist==1.5.0
google-api-core==2.24.2
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
httplib2==0.22.0
idna==3.10
multidict==6.1.0
propcache==0.3.0
proto-plus==1.26.1
protobuf==6.30.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyNaCl==1.5.0
pyparsing==3.2.1
python-dotenv==1.0.1
pytubefix==8.12.2
requests==2.32.3
rsa==4.9
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.18.3
